functions/feedback-receiver/shared/requirements-dev.txt
pytest==8.3.4
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.21.0
PyJWT==2.8.0
cryptography==42.0.4
httpx[http2]==0.27.0
flask==3.0.0
